TypeChain.ChatModels.Types

Documentation

type TypeChain model = StateT model IO

type TypeChainT = StateT

type ApiKey = ByteString

data Role

Way of distinguising who said what in a conversation

Constructors

User

Assistant

System

data Message

A message with a role and content (lenses role and content)

Constructors

Message

Fields

_role :: Role

_content :: String

role :: Lens' Message Role

content :: Lens' Message String

pattern UserMessage :: String -> Message

Pattern synonym for creating a Message with User role

pattern AssistantMessage :: String -> Message

Pattern synonym for creating a Message with Assistant role

pattern SystemMessage :: String -> Message

Pattern synonym for creating a Message with System role

class MsgList a where

Methods

toMsgList :: a -> [Message]

class ChatModel a where

A class for Chat Models

Minimal complete definition

predicts, addMsgsTo

Methods

predict :: (MonadIO m, MonadThrow m, MsgList msg) => msg -> StateT a m [Message]

Predict for current and only model

predicts :: (MonadIO m, MonadThrow m, MsgList msg) => Lens' s a -> msg -> StateT s m [Message]

Predict for a specific model via lens

addMsgs :: (Monad m, MsgList msg) => msg -> StateT a m ()

Add Messages for current and only model

addMsgsTo :: (Monad m, MsgList msg) => Lens' s a -> msg -> StateT s m ()

Add Messages for specific model

module Control.Monad.State