TypeChain.ChatModels.Types

Documentation

type TypeChain model = StateT model IO

type TypeChainT = StateT

type ApiKey = ByteString

data Role

Way of distinguising who said what in a conversation

Constructors

User

Assistant

System

data Message

A message with a role and content (lenses role and content)

Constructors

Message

Fields

_role :: Role

_content :: String

role :: Lens' Message Role

content :: Lens' Message String

pattern UserMessage :: String -> Message

Pattern synonym for creating a Message with User role

pattern AssistantMessage :: String -> Message

Pattern synonym for creating a Message with Assistant role

pattern SystemMessage :: String -> Message

Pattern synonym for creating a Message with System role

class MsgList a where

Helper typeclass to allow for multiple types to be passed to the ChatModel functions.

NOTE: If this is used with the OverloadedStrings extension, you will need type annotations when using the String instance.

Methods

toMsgList :: a -> [Message]

Convert to a list of messages

class ChatModel a where

A class for Chat Models In order to achieve compatibility with as many different kinds of LLMS as possible, the predict function is constrained to MonadIO so that it has the capability to either make an API call, run a local model, or any other action that may require IO.

Computations with a ChatModel are expected to be run in a StateT monad (see TypeChain and TypeChainT for specific types) so that the model can be updated with new messages and the output messages can be logged.

Functions that operate in a context where multiple models are available (e.g. predicts and addMsgsTo) use lenses to allow extraction and modification of the model without knowing the specific state type.

Exmaple: If working with two models, you can use (model1, model2) as the state type and pass the _1 and _2 lenses to predicts and addMsgsTo to specify which model to use in the function.

Minimal complete definition

predicts

Methods

predict :: (MonadIO m, MonadThrow m, MsgList msg) => msg -> TypeChainT a m [Message]

Predict for current and only model This function should prompt the model (either via API or locally), and return the response.

NOTE: If a model has the capability to remember previous messages, it should implement RememberingChatModel and automatically manage this functionality in the predict function.

predicts :: (MonadIO m, MonadThrow m, MsgList msg) => Lens' s a -> msg -> TypeChainT s m [Message]

Predict for a specific model via lens This function should prompt the model (either via API or locally), log the input messages, log the output messages, and return the output messages.

NOTE: If a model has the capability to remember previous messages, it should implement RememberingChatModel and automatically manage this functionality in the predicts function.

class ChatModel a => RememberingChatModel a where

Minimal complete definition

setMemoryEnabledFor, forgetFor, memorizes, rememberFor

Methods

setMemoryEnabled :: Monad m => Bool -> TypeChainT a m ()

Enable/Disable memory for current and only model

setMemoryEnabledFor :: Monad m => Lens' s a -> Bool -> TypeChainT s m ()

Enable/Disable memory for specific model

forget :: Monad m => TypeChainT a m ()

Remove all remembered messages for the current and only model. This does not affect a model's ability to remember future messages.

forgetFor :: Monad m => Lens' s a -> TypeChainT s m ()

Remove all remebered messages for a specific model. This does not affect a model's ability to remember future messages.

memorize :: Monad m => [Message] -> TypeChainT a m ()

Remember a list of messages for the current and only model. This does not affect a model's ability to remember future messages and should respect the current memory setting.

memorizes :: Monad m => Lens' s a -> [Message] -> TypeChainT s m ()

Remember a list of messages for a specific model. This does not affect a model's ability to remember future messages and should respect the current memory setting.

remember :: Monad m => TypeChainT a m [Message]

Retrieve all remembered messages for the current and only model. This does not forget any messages nor affect a model's ability to remember future messages.

rememberFor :: Monad m => Lens' s a -> TypeChainT s m [Message]

Retrieve all remembered messages for a specific model. This does not forget any messages nor affect a model's ability to remember future messages.

module Control.Monad.State